+++ server/src/modes/script/childComponents.ts
@@ -1,6 +1,6 @@
 import * as ts from 'typescript';
 import { Range } from 'vscode-languageserver-types';
-import { getLastChild, buildDocumentation, getObjectLiteralExprFromExportExpr } from './componentInfo';
+import { getLastChild, buildDocumentation, getDefaultExportObjectLiteralExpr } from './componentInfo';
 import { T_TypeScript } from '../../services/dependencyService';
 
 interface InternalChildComponent {
@@ -58,7 +58,10 @@ export function getChildComponents(
       if (objectLiteralSymbol.flags & tsModule.SymbolFlags.Alias) {
         const definitionObjectLiteralSymbol = checker.getAliasedSymbol(objectLiteralSymbol);
         if (definitionObjectLiteralSymbol.valueDeclaration) {
-          const defaultExportExpr = getLastChild(definitionObjectLiteralSymbol.valueDeclaration);
+          const defaultExportExpr = getDefaultExportObjectLiteralExpr(
+            tsModule,
+            definitionObjectLiteralSymbol.valueDeclaration.getSourceFile()
+          );
           if (!defaultExportExpr) {
             return;
           }
@@ -78,7 +81,7 @@ export function getChildComponents(
               path: definitionObjectLiteralSymbol.valueDeclaration.getSourceFile().fileName,
               range: definitionRange
             },
-            defaultExportExpr: getObjectLiteralExprFromExportExpr(tsModule, defaultExportExpr)
+            defaultExportExpr
           });
         }
       }