+++ server/src/services/typescriptService/preprocess.ts
@@ -22,7 +22,7 @@ export function parseVueScript(text: string): string {
   const doc = TextDocument.create('test://test/test.vue', 'vue', 0, text);
   const regions = getVueDocumentRegions(doc);
   const script = regions.getSingleTypeDocument('script');
-  return script.getText() || 'export default {};';
+  return script.getText();
 }
 
 function parseVueScriptSrc(text: string): string | undefined {