+++ test/interpolation/features/diagnostics/basic.test.ts
@@ -1,13 +1,8 @@
 import * as vscode from 'vscode';
-import { activateLS, showFile } from '../../helper';
 import { getDocUri, sameLineRange } from '../../util';
 import { testDiagnostics, testNoDiagnostics } from './helper';
 
 describe('Should find template-diagnostics in <template> region', () => {
-  before('activate', async () => {
-    await activateLS();
-  });
-
   const tests: TemplateDiagnosticTest[] = [
     {
       file: 'expression.vue',
@@ -180,16 +175,6 @@ describe('Should find template-diagnostics in <template> region', () => {
       ]
     },
     {
-      file: 'template-position.vue',
-      diagnostics: [
-        {
-          range: sameLineRange(13, 18, 21),
-          severity: vscode.DiagnosticSeverity.Error,
-          message: "Property 'foo' does not exist on type"
-        }
-      ]
-    },
-    {
       file: 'jsdocs-type-check.vue',
       diagnostics: [
         {
@@ -244,7 +229,30 @@ describe('Should find template-diagnostics in <template> region', () => {
   tests.forEach(t => {
     it(`Shows template diagnostics for ${t.file}`, async () => {
       const docUri = getDocUri(`diagnostics/${t.file}`);
-      await showFile(docUri);
+      await testDiagnostics(docUri, t.diagnostics, !!t.skipSameDiagnosticCountAssert);
+    });
+  });
+
+  /**
+   * These tests are somewhat flakey, especially on Windows
+   */
+  const flakeyTests: TemplateDiagnosticTest[] = [
+    {
+      file: 'template-position.vue',
+      diagnostics: [
+        {
+          range: sameLineRange(13, 18, 21),
+          severity: vscode.DiagnosticSeverity.Error,
+          message: "Property 'foo' does not exist on type"
+        }
+      ]
+    }
+  ];
+  flakeyTests.forEach(t => {
+    it(`Shows template diagnostics for ${t.file}`, async function () {
+      // Retry on fail
+      this.retries(3);
+      const docUri = getDocUri(`diagnostics/${t.file}`);
       await testDiagnostics(docUri, t.diagnostics, !!t.skipSameDiagnosticCountAssert);
     });
   });
@@ -254,7 +262,6 @@ describe('Should find template-diagnostics in <template> region', () => {
   noErrorTests.forEach(t => {
     it(`Shows no template diagnostics error for ${t}`, async () => {
       const docUri = getDocUri(`diagnostics/${t}`);
-      await showFile(docUri);
       await testNoDiagnostics(docUri);
     });
   });