+++ test/lsp/features/codeAction/basic.test.ts
@@ -1,16 +1,11 @@
 import * as vscode from 'vscode';
 import * as assert from 'assert';
-import { activateLS, showFile, getDiagnosticsAndTimeout } from '../../helper';
+import { showFile, getDiagnosticsAndTimeout } from '../../helper';
 import { getDocUri, sameLineRange } from '../../util';
-import { CodeAction, WorkspaceEdit } from 'vscode-languageclient';
+import { CodeAction } from 'vscode-languageclient';
 
 describe('Should do codeAction', () => {
   const docUri = getDocUri('codeAction/Basic.vue');
-
-  before('activate', async () => {
-    await activateLS();
-    await showFile(docUri);
-  });
 
   it('finds codeAction for unused import', async () => {
     const codeActions: CodeAction[] = [{ title: `Remove unused declaration for: 'lodash'` }];
@@ -24,6 +19,8 @@ describe('Should do codeAction', () => {
 });
 
 async function testCodeAction(docUri: vscode.Uri, range: vscode.Range, expectedActions: CodeAction[]) {
+  await showFile(docUri);
+
   await getDiagnosticsAndTimeout(docUri);
 
   const result = (await vscode.commands.executeCommand(