+++ test/lsp/features/references/basic.test.ts
@@ -1,15 +1,10 @@
-import * as vscode from 'vscode';
 import * as assert from 'assert';
-import { activateLS, showFile } from '../../helper';
-import { position, location, sameLineLocation, getDocUri } from '../../util';
+import * as vscode from 'vscode';
+import { showFile } from '../../helper';
+import { getDocUri, location, position, sameLineLocation } from '../../util';
 
 describe('Should find references', () => {
   const docUri = getDocUri('references/Basic.vue');
-
-  before('activate', async () => {
-    await activateLS();
-    await showFile(docUri);
-  });
 
   it('finds references for this.msg', async () => {
     await testReferences(docUri, position(33, 23), [