+++ test/vue3/features/completion/basic.test.ts
@@ -1,15 +1,9 @@
-import { activateLS, showFile } from '../../../lsp/helper';
-import { position, getDocUri } from '../../util';
-import { testCompletion } from './helper';
 import { CompletionItemKind } from 'vscode';
+import { getDocUri, position } from '../../util';
+import { testCompletion } from './helper';
 
 describe('Vue 3 integration test', () => {
   const fileUri = getDocUri('completion/Basic.vue');
-
-  before('activate', async () => {
-    await activateLS();
-    await showFile(fileUri);
-  });
 
   describe('Should complete Vue 3 options', () => {
     it('complete `setup`', async () => {